# enable mvstore compression: 3x storage space for 5% performance