--- a/core/src/main/java/overflowdb/storage/OdbStorage.java
+++ b/core/src/main/java/overflowdb/storage/OdbStorage.java
@@ -139,6 +139,7 @@ public class OdbStorage implements AutoCloseable {
     final MVStore store = new MVStore.Builder()
         .fileName(mvstoreFile.getAbsolutePath())
         .autoCommitBufferSize(1024 * 8)
+        .compress()
         .open();
     return store;
   }