* rough comparison of the different options with cpg2sp time and sp.bin file size for openxava:

```
no compression (master): 298M, 122s
  + manual zip compression: 48M, +5s

mvstore compress: 99M, 126s
  + manual zip compression: 65M, +5s

mvstore compressHigh: 69M, 131s
  + manual zip compression: 67M, +5s
```